+++ R/G_swOut.R
@@ -17,11 +17,8 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ###############################################################################
 
-
 # Author: Ryan J. Murphy (2013); Daniel R Schlaepfer (2013-2018)
 ###############################################################################
-
-
 
 #######
 #' Class \code{"swOUT_key"}
@@ -58,13 +55,13 @@ setClass(
   ),
   # TODO: lengths must be rSW2_glovars[["kSOILWAT2"]][["kINT"]][["SW_OUTNKEYS"]]
   prototype = list(
-    mykey = rep(NA_integer_, 34L),
-    myobj = rep(NA_integer_, 34L),
-    sumtype = rep(NA_integer_, 34L),
-    use = rep(NA, 34L),
-    first_orig = rep(NA_integer_, 34L),
-    last_orig = rep(NA_integer_, 34L),
-    outfile = rep(NA_character_, 34L)
+    mykey = rep(NA_integer_, 35L),
+    myobj = rep(NA_integer_, 35L),
+    sumtype = rep(NA_integer_, 35L),
+    use = rep(NA, 35L),
+    first_orig = rep(NA_integer_, 35L),
+    last_orig = rep(NA_integer_, 35L),
+    outfile = rep(NA_character_, 35L)
   )
 )
 
@@ -82,6 +79,16 @@ setValidity(
         names(temp)[id],
         " must be a vector of length 'SW_OUTNKEYS'"
       )
+      val <- if (isTRUE(val)) msg else c(val, msg)
+    }
+
+    if (!all(object@first_orig %in% c(NA, 1L))) {
+      msg <- "@first_orig must be 1 or NA"
+      val <- if (isTRUE(val)) msg else c(val, msg)
+    }
+
+    if (!all(object@last_orig %in% c(NA, 365, 366))) {
+      msg <- "@last_orig must be 365, 366 or NA"
       val <- if (isTRUE(val)) msg else c(val, msg)
     }
 
@@ -175,7 +182,7 @@ setClass(
     #   * 999 must be rSW2_glovars[["kSOILWAT2"]][["kINT"]][["eSW_NoTime"]]
     #   * nrows = rSW2_glovars[["kSOILWAT2"]][["kINT"]][["SW_OUTNKEYS"]]
     #   * ncols = rSW2_glovars[["kSOILWAT2"]][["kINT"]][["SW_OUTNPERIODS"]]
-    timeSteps = array(999, dim = c(34L, 4L))
+    timeSteps = array(999, dim = c(35L, 4L))
   )
 )
 
@@ -285,7 +292,9 @@ setMethod(
       # v520: `"FROZEN"` added as `outkey` 28 for a new total of 32
       to_v520 = n_has <= 31L && n_exp >= 32L,
       # v640: `"DERIVEDSUM"` and `"DERIVEDAVG"` for a new total of 34
-      to_v640 = n_has <= 32L && n_exp >= 34L
+      to_v640 = n_has <= 32L && n_exp >= 34L,
+      # v660: `"ENERGYAVG"` added as `outkey` 35 for a new total of 35
+      to_v660 = n_has <= 34L && n_exp >= 35L
     )
 
     do_upgrade <- do_upgrade[do_upgrade]
@@ -294,10 +303,8 @@ setMethod(
       target <- swOUT()
       stopifnot(nrow(target) == n_exp)
 
-
       #--- Loop over upgrades sequentially
       for (k in seq_along(do_upgrade)) {
-
         if (verbose) {
           message(
             "Upgrading object of class `swOUT`: ",
@@ -316,13 +323,18 @@ setMethod(
           to_v520 = 28L,
           # v640: `"DERIVEDSUM"` and `"DERIVEDAVG"` for a new total of 34
           to_v640 = 33L:34L,
+          # v660: `"ENERGYAVG"` added as `outkey` 35 for a new total of 35
+          to_v660 = 35L,
+          # nolint start: unreachable_code_linter.
+          # bug in lintr v3.4.0: https://github.com/r-lib/lintr/issues/3084
           stop(
-            "Upgrade ", shQuote(names(do_upgrade)[[k]]),
+            "Upgrade ",
+            shQuote(names(do_upgrade)[[k]]),
             " is not implemented for class `swOUT`.",
             call. = FALSE
           )
+          # nolint end: unreachable_code_linter.
         )
-
 
         #--- Upgrade `timeSteps`
         tmp <- object@timeSteps
@@ -356,7 +368,12 @@ setMethod(
         object@mykey <- target@mykey
 
         list_keys <- c(
-          "myobj", "sumtype", "use", "first_orig", "last_orig", "outfile"
+          "myobj",
+          "sumtype",
+          "use",
+          "first_orig",
+          "last_orig",
+          "outfile"
         )
 
         for (sn in list_keys) {
@@ -506,7 +523,6 @@ setReplaceMethod(
     object
   }
 )
-
 
 
 #' @rdname swOUT-class