+++ src/rSW_VegProd.c
@@ -41,23 +41,25 @@ vegetation production parameter information.
 /*                  Local Variables                    */
 /* --------------------------------------------------- */
 static char *cVegProd_names[] = {
-	"veg_method",
-	"nYearsDynamicShort",
-	"nYearsDynamicLong",
-	"Composition", 
-    "Albedo", 
+    "veg_method",
+    "nYearsDynamicShort",
+    "nYearsDynamicLong",
+    "Composition",
+    "Albedo",
+    "kExtVegAlbedo",
     "CanopyHeight",
-	"VegetationInterceptionParameters", 
+    "VegetationInterceptionParameters",
     "LitterInterceptionParameters",
-	"EsTpartitioning_param", 
-    "Es_param_limit", 
-    "Shade", 
+    "EsTpartitioning_param",
+    "Es_param_limit",
+    "Shade",
+    "RootProfileParameters",
     "HydraulicRedistribution_use",
-	"HydraulicRedistribution", 
-    "CriticalSoilWaterPotential", 
+    "HydraulicRedistribution",
+    "CriticalSoilWaterPotential",
     "MonthlyVeg",
-	"CO2Coefficients", 
-    "vegYear", 
+    "CO2Coefficients",
+    "vegYear",
     "isBiomAsIf100Cover"
 };
 
@@ -84,6 +86,8 @@ SEXP onGet_SW_VPD(void) {
     SEXP nYearsDynamicShort;
 	SEXP nYearsDynamicLong;
 
+    SEXP kExtVegAlbedo;
+
     SEXP Canopy, Canopy_names, Canopy_names_x;
     char *cCanopy_names_x[] = { "xinflec", "yinflec", "range", "slope", "height_cm" };
     double *p_Canopy;
@@ -102,6 +106,10 @@ SEXP onGet_SW_VPD(void) {
     SEXP Shade, Shade_names, Shade_names_x;
     char *cShade_names_x[] = { "ShadeScale", "ShadeMaximalDeadBiomass", "tanfuncXinflec", "yinflec", "range", "slope" };
     double *p_Shade;
+
+    SEXP RootProfileParameters, RootProfileParameters_names, RootProfileParameters_names_x;
+    char *cRootProfileParameters_names_x[] = { "RootShape1", "RootShape2", "MaxRootDepth" };
+    double *p_RootProfileParameters;
 
     SEXP Hydraulic_flag;//"Flag"
     SEXP Hydraulic, Hydraulic_names, Hydraulic_names_x;
@@ -165,6 +173,14 @@ SEXP onGet_SW_VPD(void) {
     }
     REAL(Albedo)[NVEGTYPES] = vi->bare_cov.albedo;
     setAttrib(Albedo, R_NamesSymbol, LandCover_names);
+
+
+    /* Get values for slot: kExtVegAlbedo */
+    PROTECT(kExtVegAlbedo = allocVector(REALSXP, NVEGTYPES));
+    for (k = 0; k < NVEGTYPES; k++) {
+        REAL(kExtVegAlbedo)[k] = vi->veg[k].kExtVegAlbedo;
+    }
+    setAttrib(kExtVegAlbedo, R_NamesSymbol, vegtype_names);
 
 
     /* Get values for slot: CanopyHeight */
@@ -255,6 +271,24 @@ SEXP onGet_SW_VPD(void) {
     SET_VECTOR_ELT(Shade_names, 0, vegtype_names);
     SET_VECTOR_ELT(Shade_names, 1, Shade_names_x);
     setAttrib(Shade, R_DimNamesSymbol, Shade_names);
+
+
+    /* Get values for slot: RootProfileParameters */
+    PROTECT(RootProfileParameters = allocMatrix(REALSXP, NVEGTYPES, 3));
+    p_RootProfileParameters = REAL(RootProfileParameters);
+    for (k = 0; k < NVEGTYPES; k++) {
+        p_RootProfileParameters[k] = vi->veg[k].rootProfileParam[0];
+        p_RootProfileParameters[k + NVEGTYPES] = vi->veg[k].rootProfileParam[1];
+        p_RootProfileParameters[k + 2 * NVEGTYPES] = vi->veg[k].rootProfileParam[2];
+    }
+    PROTECT(RootProfileParameters_names = allocVector(VECSXP, 2));
+    PROTECT(RootProfileParameters_names_x = allocVector(STRSXP, 3));
+    for (i = 0; i < 3; i++) {
+        SET_STRING_ELT(RootProfileParameters_names_x, i, mkChar(cRootProfileParameters_names_x[i]));
+    }
+    SET_VECTOR_ELT(RootProfileParameters_names, 0, vegtype_names);
+    SET_VECTOR_ELT(RootProfileParameters_names, 1, RootProfileParameters_names_x);
+    setAttrib(RootProfileParameters, R_DimNamesSymbol, RootProfileParameters_names);
 
 
     /* Get values for slot: HydraulicRedistribution_use */
@@ -360,22 +394,24 @@ SEXP onGet_SW_VPD(void) {
     SET_SLOT(VegProd, install(cVegProd_names[2]), nYearsDynamicLong);
     SET_SLOT(VegProd, install(cVegProd_names[3]), VegComp);
     SET_SLOT(VegProd, install(cVegProd_names[4]), Albedo);
-    SET_SLOT(VegProd, install(cVegProd_names[5]), Canopy);
-    SET_SLOT(VegProd, install(cVegProd_names[6]), VegInterception);
-    SET_SLOT(VegProd, install(cVegProd_names[7]), LitterInterception);
-    SET_SLOT(VegProd, install(cVegProd_names[8]), EsTpartitioning_param);
-    SET_SLOT(VegProd, install(cVegProd_names[9]), Es_param_limit);
-    SET_SLOT(VegProd, install(cVegProd_names[10]), Shade);
-    SET_SLOT(VegProd, install(cVegProd_names[11]), Hydraulic_flag);
-    SET_SLOT(VegProd, install(cVegProd_names[12]), Hydraulic);
-    SET_SLOT(VegProd, install(cVegProd_names[13]), CSWP);
-    SET_SLOT(VegProd, install(cVegProd_names[14]), MonthlyVeg);
-    SET_SLOT(VegProd, install(cVegProd_names[15]), CO2Coefficients);
-    SET_SLOT(VegProd, install(cVegProd_names[16]), VegYear);
-    SET_SLOT(VegProd, install(cVegProd_names[17]), IsBiomAsIf100Cover);
+    SET_SLOT(VegProd, install(cVegProd_names[5]), kExtVegAlbedo);
+    SET_SLOT(VegProd, install(cVegProd_names[6]), Canopy);
+    SET_SLOT(VegProd, install(cVegProd_names[7]), VegInterception);
+    SET_SLOT(VegProd, install(cVegProd_names[8]), LitterInterception);
+    SET_SLOT(VegProd, install(cVegProd_names[9]), EsTpartitioning_param);
+    SET_SLOT(VegProd, install(cVegProd_names[10]), Es_param_limit);
+    SET_SLOT(VegProd, install(cVegProd_names[11]), Shade);
+    SET_SLOT(VegProd, install(cVegProd_names[12]), RootProfileParameters);
+    SET_SLOT(VegProd, install(cVegProd_names[13]), Hydraulic_flag);
+    SET_SLOT(VegProd, install(cVegProd_names[14]), Hydraulic);
+    SET_SLOT(VegProd, install(cVegProd_names[15]), CSWP);
+    SET_SLOT(VegProd, install(cVegProd_names[16]), MonthlyVeg);
+    SET_SLOT(VegProd, install(cVegProd_names[17]), CO2Coefficients);
+    SET_SLOT(VegProd, install(cVegProd_names[18]), VegYear);
+    SET_SLOT(VegProd, install(cVegProd_names[19]), IsBiomAsIf100Cover);
 
     /* Memory clean up */
-    UNPROTECT(37);
+    UNPROTECT(41);
 
     return VegProd;
 }
@@ -391,6 +427,7 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
     SEXP nYearsDynamicShort;
     SEXP nYearsDynamicLong;
     SEXP Albedo;
+    SEXP kExtVegAlbedo;
     SEXP Canopy;
     double *p_Canopy;
     SEXP VegInterception;
@@ -401,6 +438,8 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
     SEXP Es_param_limit;
     SEXP Shade;
     double *p_Shade;
+    SEXP RootProfileParameters;
+    double *p_RootProfileParameters;
     SEXP Hydraulic;
     SEXP Hydraulic_flag;
     SEXP CSWP;
@@ -441,8 +480,15 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
     vi->bare_cov.albedo = REAL(Albedo)[NVEGTYPES]; //Bare Ground
 
 
+    /* Set values using slot: kExtVegAlbedo */
+    PROTECT(kExtVegAlbedo = GET_SLOT(SW_VPD, install(cVegProd_names[5])));
+    for (k = 0; k < NVEGTYPES; k++) {
+        vi->veg[k].kExtVegAlbedo = REAL(kExtVegAlbedo)[k];
+    }
+
+
     /* Set values using slot: CanopyHeight */
-    PROTECT(Canopy = GET_SLOT(SW_VPD, install(cVegProd_names[5])));
+    PROTECT(Canopy = GET_SLOT(SW_VPD, install(cVegProd_names[6])));
     p_Canopy = REAL(Canopy);
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].cnpy.xinflec = p_Canopy[k];
@@ -454,7 +500,7 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
 
 
     /* Set values using slot: VegetationInterceptionParameters */
-    PROTECT(VegInterception = GET_SLOT(SW_VPD, install(cVegProd_names[6])));
+    PROTECT(VegInterception = GET_SLOT(SW_VPD, install(cVegProd_names[7])));
     p_VegInterception = REAL(VegInterception);
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].veg_kSmax = p_VegInterception[k];
@@ -463,7 +509,7 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
 
 
     /* Set values using slot: LitterInterceptionParameters */
-    PROTECT(LitterInterception = GET_SLOT(SW_VPD, install(cVegProd_names[7])));
+    PROTECT(LitterInterception = GET_SLOT(SW_VPD, install(cVegProd_names[8])));
     p_LitterInterception = REAL(LitterInterception);
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].lit_kSmax = p_LitterInterception[k];
@@ -471,21 +517,21 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
 
 
     /* Set values using slot: EsTpartitioning_param */
-    PROTECT(EsTpartitioning_param = GET_SLOT(SW_VPD, install(cVegProd_names[8])));
+    PROTECT(EsTpartitioning_param = GET_SLOT(SW_VPD, install(cVegProd_names[9])));
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].EsTpartitioning_param = REAL(EsTpartitioning_param)[k];
     }
 
 
     /* Set values using slot: Es_param_limit */
-    PROTECT(Es_param_limit = GET_SLOT(SW_VPD, install(cVegProd_names[9])));
+    PROTECT(Es_param_limit = GET_SLOT(SW_VPD, install(cVegProd_names[10])));
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].Es_param_limit = REAL(Es_param_limit)[k];
     }
 
 
     /* Set values using slot: Shade */
-    PROTECT(Shade = GET_SLOT(SW_VPD, install(cVegProd_names[10])));
+    PROTECT(Shade = GET_SLOT(SW_VPD, install(cVegProd_names[11])));
     p_Shade = REAL(Shade);
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].shade_scale = p_Shade[k];
@@ -497,15 +543,25 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
     }
 
 
+    /* Set values using slot: RootProfileParameters */
+    PROTECT(RootProfileParameters = GET_SLOT(SW_VPD, install(cVegProd_names[12])));
+    p_RootProfileParameters = REAL(RootProfileParameters);
+    for (k = 0; k < NVEGTYPES; k++) {
+        vi->veg[k].rootProfileParam[0] = p_RootProfileParameters[k];
+        vi->veg[k].rootProfileParam[1] = p_RootProfileParameters[k + NVEGTYPES];
+        vi->veg[k].rootProfileParam[2] = p_RootProfileParameters[k + 2 * NVEGTYPES];
+    }
+
+
     /* Set values using slot: HydraulicRedistribution_use */
-    PROTECT(Hydraulic_flag = GET_SLOT(SW_VPD, install(cVegProd_names[11])));
+    PROTECT(Hydraulic_flag = GET_SLOT(SW_VPD, install(cVegProd_names[13])));
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].flagHydraulicRedistribution = LOGICAL_POINTER(Hydraulic_flag)[k];
     }
 
 
     /* Set values using slot: HydraulicRedistribution */
-    PROTECT(Hydraulic = GET_SLOT(SW_VPD, install(cVegProd_names[12])));
+    PROTECT(Hydraulic = GET_SLOT(SW_VPD, install(cVegProd_names[14])));
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].maxCondroot = REAL(Hydraulic)[k];
         vi->veg[k].swpMatric50 = REAL(Hydraulic)[k + NVEGTYPES];
@@ -514,7 +570,7 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
 
 
     /* Set values using slot: CriticalSoilWaterPotential */
-    PROTECT(CSWP = GET_SLOT(SW_VPD, install(cVegProd_names[13])));
+    PROTECT(CSWP = GET_SLOT(SW_VPD, install(cVegProd_names[15])));
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].SWPcrit = -10 * REAL(CSWP)[k];
         // for use with get_swa for properly partitioning swa
@@ -524,7 +580,7 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
 
 
     /* Set values using slot: MonthlyVeg */
-    PROTECT(MonthlyVeg = GET_SLOT(SW_VPD, install(cVegProd_names[14])));
+    PROTECT(MonthlyVeg = GET_SLOT(SW_VPD, install(cVegProd_names[16])));
     for (k = 0; k < NVEGTYPES; k++) {
         PROTECT(monBiomass =  VECTOR_ELT(MonthlyVeg, k));
         p_monBiomass = REAL(monBiomass);
@@ -539,7 +595,7 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
 
 
     /* Set values using slot: CO2Coefficients */
-    PROTECT(CO2Coefficients = GET_SLOT(SW_VPD, install(cVegProd_names[15])));
+    PROTECT(CO2Coefficients = GET_SLOT(SW_VPD, install(cVegProd_names[17])));
     for (k = 0; k < NVEGTYPES; k++) {
         vi->veg[k].co2_bio_coeff1 = REAL(CO2Coefficients)[k];
         vi->veg[k].co2_bio_coeff2 = REAL(CO2Coefficients)[k + NVEGTYPES];
@@ -549,17 +605,24 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
 
 
     /* Set values using slot: vegYear */
-    PROTECT(VegYear = GET_SLOT(SW_VPD, install(cVegProd_names[16])));
+    PROTECT(VegYear = GET_SLOT(SW_VPD, install(cVegProd_names[18])));
     vi->vegYear = INTEGER(VegYear)[0];
 
 
     /* Set values using slot: isBiomAsIf100Cover */
-    PROTECT(IsBiomAsIf100Cover = GET_SLOT(SW_VPD, install(cVegProd_names[17])));
+    PROTECT(IsBiomAsIf100Cover = GET_SLOT(SW_VPD, install(cVegProd_names[19])));
     vi->isBiomAsIf100Cover = LOGICAL(IsBiomAsIf100Cover)[0];
 
 
     /* Wrap up */
-    SW_VPD_fix_cover(&SoilWatRun.RunIn.VegProdRunIn, LogInfo);
+    fixVegCoverInputs(&SoilWatRun.RunIn.VegProdRunIn, LogInfo);
+    if (LogInfo->stopRun) {
+        goto freeMem; // Exit function prematurely due to error
+    }
+
+    checkVegetationInputs(
+        &SoilWatRun.VegProdIn, &SoilWatRun.RunIn.VegProdRunIn, LogInfo
+    );
     if (LogInfo->stopRun) {
         goto freeMem; // Exit function prematurely due to error
     }
@@ -569,7 +632,7 @@ void onSet_SW_VPD(SEXP SW_VPD, LOG_INFO* LogInfo) {
     }
 
 freeMem:
-    UNPROTECT(18);
+    UNPROTECT(20);
 }
 
 // `estimate_PotNatVeg_composition()` is R interface to rSW2_estimate_PotNatVeg_composition()